Fix t-lsymbol and ja-anthy test cases
If these do not set no dictionary, there might be candidates from
dictionaries when no candidates from m17n are expected because the
default dictionaries depend on the locale. And that means that
depending on the locale the test is run in and the specific
dictionaries installed on the system (and the configuration of the
spellchecker enchant) a preedit content of something like “a☺️b”
may surprisingly produce candidates.