Reduce match_limit in emoji-picker by default to 1_000 again but add a --match-limit command line option
A default match limit of 100_000 is too big after all. The search itself is fast enough, but if it happens
that many thousands characters match the rendering of all these glyphs in the flowbox may take “forever”.

Usually 1_000 should be plenty and if one really needs more there is the --match-limit command line option now.