Show the code point and a comment in the lookup table for control characters or other invisible characters
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/715

Do not put these characters into the lookup table directly, that may cause problems,
just show the code point and a comment instead.

Selecting them still works because the string which is committed is not read
from the lookup table but from `self._candidates[index].phrase`.

For a candidate text in the lookup table which is longer than just one character but
consists **only** out of invisible characters, show the Python representagion of that text
(`repr(text)`).

Show a comment in the lookup table for control characters or other invisible characters