When committing with space while the preedit is empty, something might be selected in the lookup table which needs to be committed then
Related: https://github.com/mike-fabian/ibus-typing-booster/issues/707

The fix for
https://github.com/mike-fabian/ibus-typing-booster/issues/707 was
incomplete: It is possible that a lookup table is shown when the input
is empty. And it is possible to select something in such a lookup
table and then commit with space.  In that case, the selection in the
lookup table needs to be committed and the space appended, it is not
OK just to discard the lookup table.