command `show_selection_info`: Instead of cancelling the selection, add the selection to the candidates
Related: https://github.com/mike-fabian/ibus-typing-booster/issues/733

Cancelling the selection with Control+c Control+v is a really ugly
hack, with David Mandelberg’s idea to add the contents of the
selection to the candidates this hack can be avoided: If one of the
candidates is committed, the still selected text is replaced by the
candidate but as every candidate is prefixed with the contents of the
selection, this immediately restores the original text.

But there was the side effect that the original text was not always
faithfully restored, it was converted to NFC because Typing Booster
converted to NFC on every commit.

To avoid that the reinserted selection gets normalized, I now stopped
normalizing on every commit.

Instead I normalize the candidates coming from
itb_util.best_candidates(), doing that the candidates coming from the
user database or the hunspell dictionaries get normalized. That is
probably enough, I think it is OK not to normalize a preedit which is
committed without selecting a candidate and it is probably also OK not
to normalize m17n candidate lists.