Move normalization to NFKD for ko-romaja and ko-han2 to _update_candidates()
The normalization to NFKD for ko-romaja and ko-han2 is needed for the
matching in /usr/share/myspell/ko_KR.dic and the user database to work
correctly.  Therefore, move it to _update_candidates(). This makes
sure that the matching still works but avoids it in other places like
the preedit.