Do not try to point the popovers on search results directly to the emoji
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/766

Trying to point directly to the emoji is difficult as it is hard to
tell where the emoji actually is.

For left-to-right languages it is on the left. For right-to-left
langauges the emoji is typically on the right but may be on the left
it the name of the emoji lacks a translation.

It is better not to use `.set_pointing_to(rectangle)` at all. Without
that it automatically points to the middle of the event box which is
OK always, no matter for which language and no matter whether a
translation of the emoji name is available or not.

But then also use

 self._emoji_info_popover.set_position(Gtk.PositionType.BOTTOM)

instead of Gtk.PositionType.RIGHT for the emoji description popover
to avoid hiding parts of the emoji name.