Use GLib.timeout_add() instead of GLib.idle_add() to force a short delay after wl-paste
If `wl-paste -p` is used to get the primary selection, it causes focus
out and focus in events.  This might close the candidate if is was
already shown or prevent it from appearing.  Using
GLib.timeout_add(100, ...) gives the focus events some time to happen
before the candidate list is shown.

Related: https://github.com/mike-fabian/ibus-typing-booster/issues/787#issuecomment-3236363564