Add <path>~/.config/dconf/</path> to <observed-paths> in the component file
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/788

The ibus registry needs to be updated when some typing booster
settings change.  I could call `ibus write-cache` after changing a
setting which needs an update of the registry, but adding the
directory where the settings are stored to observed-paths works just
as well. It seems simpler and it also documents that the registry
might need an update if a setting has changed.

`~/.m17n.d` is already in observed paths and it if that exists, the
problem does not occur. Any directory in observed paths which is an
existing directory in the users home directory seems to trigger an
update of the registry when doing `ibus restart`. But `~/.m17n.d` does
need to exist and instead of creating it is better to use
`~/.config/dconf` because this surely always exists and better shows the
intention.

Thanks to David Mandelberg <david@mandelberg.org> for the idea to use
`~/.config/dconf` instead of `~/.m17n.d`.