Call `ibus write-cache` when the inputmodetruesymbol is changed or out of date
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/788
Related: https://github.com/mike-fabian/ibus-typing-booster/issues/792

When the inputmodetruesymbol is changed manually using the setup tool
or the command line, the ibus registry cache needs to be updated.
Call `ibus write-cache` directly in that case to make sure the cache
really gets updated.

When Typing Booster starts, call `ibus read-cache` to check whether
the desired symbol is already in the cache. If it is not in the cache,
call `ibus write-cache` to force an update of the cache. This should
make the symbol correct after another `ibus restart` or desktop restart.

Putting `~/.config/dbus/` into `observed-paths` to get automatic
updates can causes slower session startups.