Add a ibus-setup-tb.desktop file
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/807

A desktop file `ibus-setup-<prefix>.desktop` is necessary to enable
`gnome-control-center keyboard` to start the setup tool of an engine
with a name identical to `<prefix>` or starting with `<prefix>:`.

For example, if the engine name is `typing-booster`, then
`gnome-control-center keyboard` will start the program specified with
`Exec=program` in `ibus-setup-typing-booster.desktop`. And if the
engine name is `tb:ar:translit` then `gnome-control-center keyboard`
will start the program specified in `ibus-setup-tb.desktop`.

The program is started without command line arguments:

$ grep Exec *desktop.in.in
ibus-setup-tb.desktop.in.in:Exec=${LIBEXECDIR}/ibus-setup-typing-booster
ibus-setup-typing-booster.desktop.in.in:Exec=${LIBEXECDIR}/ibus-setup-typing-booster

But `gnome-control-center keyboard` sets the environment variable
'IBUS_ENGINE_NAME' to the engine name.  Therefore, the setup program
needs to check this environment variable if the command line option
`--engine-name` is not used.