Improve behaviour on OSK for “Enable suggestions by key” and “Minimum number of chars for completion“ options
Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/813
([BUG] "☑️ Enable suggestions by key" should not prevent predictions on OSK)

Resolves: https://github.com/mike-fabian/ibus-typing-booster/issues/814
([BUG] The setting "Minimum number of chars for completion" should be ignored on OSK)

It also improves a bit the behaviour when “Minimum number of chars for
completion“ is 0 *and* OSK is not used: When there are problems
getting the correct surrounding text I don’t try to show predictions
on empty input at all anymore, no fallback to remembered context is
tried to replace the missing or wrong surrounding text (Unless when
doing unit testing, as surrounding text does not work during unit
testing, remembered context is used to be able to do a test at all).

That means that there will be fewer situations when suggestions on
empty input can be shown. Better show fewer but good suggestions than
showing too many garbage suggestions.

If OSK is used, suggestions on empty input are now shown when updating
the surrounding text after the last commit works correctly. Using

    self._trigger_surrounding_text_update()
    # Tiny delay of to give the surrounding text a chance to update:
    GLib.timeout_add(5, self._update_ui_empty_input_try_completion)

after commits is necessary to get up-to-date surrounding text when
self._update_ui_empty_input_try_completion() is executed.