Add `class TypingBoosterAuxiliaryText` and `class TypingBoosterPreeditText`
Makes tracking content and visibility of the auxiliary text and the preedit text easier.

IBus.Engine has no methods to query the current contents of the
auxiliary text and whether it is visible or not.

To make it easier to track the current contents and visibility of the
auxiliary text I added a `class TypingBoosterAuxiliaryText`
and did override

IBus.Engine.show_auxiliary_text()
IBus.Engine.hide_auxiliary_text()
IBus.Engine.update_auxiliary_text(text: IBus.Text, visbile: bool)

with methods which track the contents and visibility.

And the same refactoring also for the preedit test.